requirements.txt
earthengine-api
rasterio
geopandas
numpy
pandas
matplotlib
torch
torchvision
pyyaml
scikit-learn